# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Legend Academy API Documentation",
        "description": "Legend Pay API description",
        "contact": {
            "email": "o.oyedele@suburbanfiberco.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get list of courses",
                "description": "Returns list of courses",
                "operationId": "getCoursesList",
                "parameters": [
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Course Retrieved Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Store new course",
                "description": "Returns course data",
                "operationId": "storeCourse",
                "parameters": [
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Course created successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Course"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/courses-modules": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get list of courses with modules",
                "description": "Returns list of courses with modules",
                "operationId": "Course Modules",
                "parameters": [
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Course Retrieved Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CourseModule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/v1/courses/{id}/modules": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get Course information",
                "description": "Returns course data",
                "operationId": "getCourseModuleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Course Module Retrieved Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Module"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/publish-course/{id}": {
            "patch": {
                "tags": [
                    "Courses"
                ],
                "summary": "Publish a course",
                "description": "Returns no course data",
                "operationId": "publishCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer_token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "message",
                                            "type": "string",
                                            "example": "Course Published Successfully"
                                        },
                                        "success": {
                                            "description": "success",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Course": {
                "title": "Course",
                "description": "Course model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the new course",
                        "type": "string",
                        "example": "Introduction to workpkace"
                    },
                    "number_of_modules": {
                        "title": "Number Of Modules",
                        "description": "Total Number Of Modules",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "number_enrolled": {
                        "title": "Number Of Enrolled",
                        "description": "Total Number Of Enrolled Student",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "User's id of the new course",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "department_id": {
                        "title": "Department ID",
                        "description": "Department's id of the new course",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "unit_id": {
                        "title": "Unit ID",
                        "description": "Department unit ID of the new course",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "status": {
                        "title": "Status",
                        "description": "Status of the new course",
                        "type": "integer",
                        "format": "int64",
                        "example": "Published"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Course"
                }
            },
...
        }
    }
}